move icons in atoms/icons
