+++ src/components/atomic/buttons/DavitAddButton.tsx
@@ -1,5 +1,5 @@
 import React, { CSSProperties, FunctionComponent } from "react";
-import { DavitIcons } from "../IconSet";
+import { DavitIcons } from "../icons/IconSet";
 import { DavitButton } from "./DavitButton";
 
 interface DavitAddButtonProps {
