+++ src/components/atomic/icons/IconActorRemove.tsx
