+++ src/components/atomic/icons/IconArrowLeft.tsx
