+++ src/components/atomic/icons/IconArrowRight.tsx
