+++ src/components/atomic/icons/IconUp.tsx
