+++ src/components/atomic/icons/IconZoomOut.tsx
