+++ src/components/atomic/icons/index.ts
