+++ src/components/exxcellent-components/atoms/index.ts
@@ -1,3 +1,3 @@
-export * from "./icons";
+export * from "../../atomic/icons";
 export * from "./bars";
 export * from "./cards";
