+++ src/components/molecules/notifications/DavitNotification.tsx
@@ -2,7 +2,7 @@ import { IconDefinition } from "@fortawesome/fontawesome-common-types";
 import { faInfoCircle } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React, { FunctionComponent } from "react";
-import { DavitIcons } from "../../atomic/IconSet";
+import { DavitIcons } from "../../atomic/icons/IconSet";
 import "./DavitNotification.css";
 
 export enum NotificationLevel {
