+++ src/components/organisms/table/DavitTable.tsx
@@ -1,7 +1,7 @@
 import { IconDefinition } from "@fortawesome/fontawesome-common-types";
 import React, { FunctionComponent } from "react";
 import { DavitTableButton } from "../../atomic/buttons/DavitTableButton";
-import { DavitIcons } from "../../atomic/IconSet";
+import { DavitIcons } from "../../atomic/icons/IconSet";
 import "./DavitTable.css";
 
 export interface DavitTableProps {
