+++ src/domains/overview/tableModel/tables/model/ModelChain.ts
@@ -1,5 +1,5 @@
 import { useDispatch } from "react-redux";
-import { DavitIcons } from "../../../../../components/atomic/IconSet";
+import { DavitIcons } from "../../../../../components/atomic/icons/IconSet";
 import { DavitTableRowData } from "../../../../../components/organisms/table/DavitTable";
 import { ChainTO } from "../../../../../dataAccess/access/to/ChainTO";
 import { EditActions } from "../../../../../slices/EditSlice";
