+++ src/domains/overview/tableModel/tables/model/ModelDataSetup.ts
@@ -1,5 +1,5 @@
 import { useDispatch } from "react-redux";
-import { DavitIcons } from "../../../../../components/atomic/IconSet";
+import { DavitIcons } from "../../../../../components/atomic/icons/IconSet";
 import { DavitTableRowData } from "../../../../../components/organisms/table/DavitTable";
 import { DataSetupTO } from "../../../../../dataAccess/access/to/DataSetupTO";
 import { EditActions } from "../../../../../slices/EditSlice";
