+++ src/domains/overview/tableModel/tables/model/ModelSequence.ts
@@ -1,5 +1,5 @@
 import { useDispatch } from "react-redux";
-import { DavitIcons } from "../../../../../components/atomic/IconSet";
+import { DavitIcons } from "../../../../../components/atomic/icons/IconSet";
 import { DavitTableAction, DavitTableRowData } from "../../../../../components/organisms/table/DavitTable";
 import { SequenceTO } from "../../../../../dataAccess/access/to/SequenceTO";
 import { EditActions } from "../../../../../slices/EditSlice";
